Highstone-30.06.2019

+++ Game/core/src/com/mygdx/game/controlling/UserData.java
@@ -5,9 +5,9 @@ import java.util.List;
 public class UserData {
 
     public String name;
-    public List<Integer> highscores;
+    public List<Long> highscores;
 
-    public UserData(String name, List<Integer> highscores) {
+    public UserData(String name, List<Long> highscores) {
         this.name = name;
         this.highscores = highscores;
     }
